requirements.txt
asgiref==3.6.0
async-generator==1.10
attrs==22.2.0
certifi==2022.12.7
charset-normalizer==3.0.1
Django==4.1.6
exceptiongroup==1.1.0
h11==0.14.0
idna==3.4
MouseInfo==0.1.3
numpy==1.24.1
outcome==1.2.0
packaging==23.0
pandas==1.5.3
PyAutoGUI==0.9.53
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyperclip==1.8.2
PyRect==0.2.0
PyScreeze==0.1.28
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==0.21.1
python3-xlib==0.15
pytweening==1.0.4
pytz==2022.7.1
requests==2.28.2
selenium==4.8.0
six==1.16.0
smsactivate==1.5
sniffio==1.3.0
sortedcontainers==2.4.0
sqlparse==0.4.3
tqdm==4.64.1
trio==0.22.0
trio-websocket==0.9.2
undetected-chromedriver==3.2.1
urllib3==1.26.14
webdriver-manager==3.8.5
websockets==10.4
wsproto==1.2.0
